refactor: move hook to file
